LayoutTests/imported/w3c/web-platform-tests/css/css-scroll-snap/scroll-initial-target/scroll-initial-target-with-scroll-snap.tentative-expected.txt


FAIL  CSS Scroll Snap 2 Test: scroll-initial-target* assert_approx_equals: scroll-initial-target sets initial horizontal scroll position expected 200 +/- 1 but got 0
